requirements.txt
altair                    5.5.0
attrs                     25.3.0
blinker                   1.9.0
branca                    0.8.1
cachetools                5.5.2
certifi                   2025.7.14
charset-normalizer        3.4.2
click                     8.2.1
click-plugins             1.1.1.2
cligj                     0.7.2
colorama                  0.4.6
fiona                     1.9.6
folium                    0.17.0
geopandas                 0.14.4
gitdb                     4.0.12
GitPython                 3.1.45
idna                      3.10
Jinja2                    3.1.6
joblib                    1.5.1
jsonschema                4.25.0
jsonschema-specifications 2025.4.1
markdown-it-py            3.0.0
MarkupSafe                3.0.2
mdurl                     0.1.2
narwhals                  2.0.1
networkx                  3.3
numpy                     1.26.4
osmnx                     1.9.4
packaging                 24.2
pandas                    2.1.4
pillow                    10.4.0
pip                       24.2
protobuf                  5.29.5
pyarrow                   21.0.0
pydeck                    0.9.1
Pygments                  2.19.2
pyproj                    3.7.1
python-dateutil           2.9.0.post0
python-dotenv             1.0.0
pytz                      2025.2
referencing               0.36.2
requests                  2.31.0
rich                      13.9.4
rpds-py                   0.26.0
scikit-learn              1.3.2
scipy                     1.16.1
shapely                   2.0.7
six                       1.17.0
smmap                     5.0.2
streamlit                 1.38.0
streamlit_folium          0.22.0
tenacity                  8.5.0
threadpoolctl             3.6.0
toml                      0.10.2
tornado                   6.5.1
typing_extensions         4.14.1
tzdata                    2025.2
urllib3                   2.5.0
watchdog                  4.0.2
xyzservices               2025.4.0